refactor: refactor code using express routers to consolidate the server file
